Fix(DistrictDashboard): Seperated some FormData props and logic to seperate file using a custom hook for cleaner code
